requirements.txt
pytest==8.1.1
pytest-selenium==4.1.0
pytest-html
allure-pytest
selenium
webdriver-manager
requests
urllib3
pydantic
pydantic-settings